expose table and array typedef in the header file
